chore: update dependencies and enhance code quality (#73)
* chore: update dependencies and enhance code quality

Upgrade Vite to version 6.2.0 and improve code maintenance by adding 
additional Prism.js components for syntax highlighting. Update 
@esbuild packages to version 0.25.0 and globals package to 16.0.0 to 
ensure compatibility and take advantage of the latest features and 
bug fixes. Add error logging in the load function and make minor 
type adjustments in the TechStack interface.

* feat: enhance HTML structure and accessibility improvements

Add a title to the application for improved SEO and user experience. 
Replace inline styles with a class for the SvelteKit body 
wrapper to maintain a clearer structure and improve readability. 
Refactor button disabling logic in the navigation to streamline 
the code and enhance maintainability.